requirements.txt
click>=8.0.0
openai>=1.0.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
python-dotenv>=0.19.0
PyYAML>=6.0.0
tqdm>=4.65.0